split out activity models
