+++ api/db/activity/approach.js
@@ -1,0 +1,22 @@
+module.exports = {
+  apdActivityApproach: {
+    tableName: 'activity_approaches',
+
+    activity() {
+      return this.belongsTo('apdActivity');
+    },
+
+    toJSON() {
+      return {
+        id: this.get('id'),
+        description: this.get('description'),
+        alternatives: this.get('alternatives'),
+        explanation: this.get('explanation')
+      };
+    },
+
+    static: {
+      updateableFields: ['description', 'alternatives', 'explanation']
+    }
+  }
+};
